调整留言板的结构，拆分 schema 文件
